// src/store/api/types.ts
export interface IAddress {
  street: string;
  street2: string;
  city: string;
  state: string;
  zipCode: string;
}
// ...
export interface IQuoteEstimate {
  type: string;
  duration: number;
  productId: string;
  countyFips: string;
  netRateOnline: number;
  premiumAmount: number;
  coverageAmount: number;
}